Cleaning Some of the Wording
Jayce,

I cleaned up some of the wording for the readME file. Have a look and let me know if you have any questions. Also, I stated that you are using dispersion in the model but am not 100% sure if dispersion is still a part of this model. Please correct if I am wrong.

Finally, I made the edits in Notepad, so please let me know if this file got messed up and send me the name of the VisualBasic package you use.

Thanks,

John